code-review: iteration 2
